Sugarcube and Space Police Housekeeping (#1654)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR updates the the Sugarcube mildly, while removing it from the
normal rotation of ships. It's unfortunate, but it'll remain spawnable
on the admin side.
It also updates and moves the hunter_space_cop ship, turning it into a
subshuttle to facilitate admin shenanigans.
![2022 12 29-13 39
19](https://user-images.githubusercontent.com/94164348/210013521-dfed2c6d-fac6-4016-aab8-0f2110db9974.png)
![2022 12 29-13 42
31](https://user-images.githubusercontent.com/94164348/210013729-6509e5d7-c2d7-4769-81c0-5774ae238ec9.png)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
One of these was going to be deprecated and the other just needed an
update.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: The Kunai Dropship, a small ship designed to make admin shenanigans
mildly easier
del: The Sugarcube is no longer enabled, ask an admin if you'd like to
have it spawned.
tweak: Sugarcube has received a minor interior facelift.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>