Nanotrasen Mimir-Class Rehabilitation (Prison) Vessel (#1724)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the Mimir-Class Rehabilitation (Prison) Vessel
Removes the Tartarus
It's meant to simulate perma and the gulag, mixed with a bunch of maint
stuff and low security staff. Starts with 3 security and 1
patient/prisoner.
![2023 04 20-14 53
43](https://user-images.githubusercontent.com/55075747/233496215-50d2bdae-8515-41d6-8be1-ab7ceaf8562e.png)
![2023 04 20-14 53
46](https://user-images.githubusercontent.com/55075747/233496220-d6708625-1bea-433b-a8b6-d7b7228ebf47.png)

![mimirreal](https://user-images.githubusercontent.com/55075747/233496222-6ae72eb0-9e5d-4604-8624-e5f10f51030a.png)



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
The tartarus was our security/perma ship, so I want something better.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Adds a mimir (sleepy)
code: adds some supporting code for the mimir (sleepy)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>