ORM QOL: input/output directions can now be changed separately (#1953)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the ability to whack an ORM with a multitool while its panel is
open to change whether or not alt+clicking it will rotate the input
direction, output direction, or both. Default functionality on a fresh
or mapped ORM is still both.

## Why It's Good For The Game

Currently ORMs can only output on the tile directly opposite the tile
they take in ore from. This worked fine back on a station, but on a ship
space is limited, and the crewmember mining might be the crewmember
needing those materials immediately afterward.

Now, shiptesters no longer have to fear ORMs spitting their contents
into walls or furniture when building one in a corner, dead end, or
other spot where having the input/output tiles be directly opposite each
other posed logistical problems.

## Changelog

:cl:
code: multitools now change the alt+click behavior of an ORM
tweak: ORMs can now have their input and output tiles onto the same
tile, or two tiles at a right angle, as well as the old behavior of
having them on opposite sides of the machine
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->