Literally removes RnD from every ship (#1958)
![as_we_speak](https://github.com/shiptest-ss13/Shiptest/assets/60533805/84feba91-d1f6-42dc-bf40-ceda42ae18bc)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes RnD from every ship, potentially as we speak.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Ships are too independent and the job of the scientist and medic is far
too stale- All they do is stand in one place and click buttons. Removing
RnD encourages cooperation and more interaction with the outpost cargo
system.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Fixes RnD balance
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->