The SolGov-ening Part 1: Clothes and The Purge (#1968)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Hello there!

This pull request is the first on *many* to come that will slowly but
surely add new content for SolGov, finally bringing it to parity with
the lore! ~~(most of which was hidden in people's brains)~~

This first PR will address outfits, adding a multitude of new ones:


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/e23249d4-712f-450f-af9e-f0e77794272b)

In a very short description, most of these outfits are related to
existing jobs, such as the SolGov Representative, but several *new* ones
now exist in implemented spritework for now:

- The Sonnensoldner (Solarian Mercenary)
- The Overseer
- The Bureaucrat
- The Civilian
- The Solarian Captain

This PR also includes minor additions, such as:

- A new Table Flag


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/b3705d6e-cec3-4db4-9b68-dc87cddbf99b)

- A new Bedsheet


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/ae8893d8-a3ba-4681-bacd-c3356d1ce275)

This PR also **removes** and edits much of the code related to SolGov.
Many things related to old SolGov, such as most marine equipment
clothing-wise has been completely excised from the game. Old job datums
that were only used for deprecated SolGov ships were also mostly
eliminated. Some might still be around but that's for future me to deal
with.

Speaking of! This PR also fully removes the Liberty-Class (lol) and the
Cricket-Class. May they rest in peace, after serving us for... A year or
two I think.

Finally, this PR adds a special new variable: `worn_y_offset`. Utilize
this var to offset any clothing sprite along the Y axis!

Special thanks to:

@triplezeta For helping me make sure the lore is accurate.

@PiperDoots For providing the original work I would eventually overhaul.

@MarkSuckerberg For the worn_y_offset code.

~~@retlaw34 For providing this poor bastard some emotional support in
VC.~~


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I really do hope that I don't have to explain why a rework 6 months in
the making finally being implemented is good for the game.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy, MarkSuckerberg, TetraZeta, PiperDoots
add: The start of new SolGov! Expect a new and improved SolGov to now
roam the frontier!
del: Removes many old jobs and content related to Old Solgov.
del: Fully deprecates the Cricket-Class.
del: Fully deprecates the Liberty-Class.
code: Adds the worn_y_offset variable, making offsetting hats much
easier.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->