Ember Class Pirate Raider (#1771)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This ship was once used for exploration when settlers first entered the
sector. Though it has been many rotations since this model last glided
through the empty vacuum of space. While most of them lay abandoned, or
in scrapyards, one of them has managed to escape this faith. This ship,
The Cutter, is now the home of a particular group of Frontiersmen. While
the ship has seen better days, its size is more than enough to strike
fear into any merchant ship unlucky enough to enter its domain.

so! let's get into what this ship is, and what it does. 
the ship is used by the Frontiersmen, as a show of force, raiding, and
ship-to-ship combat. on the front, it has tw- one gun bay, capable of
launching explosives (or people) at whatever foolishly somehow places
itself in front of the barrel. it has a decently stocked armory, a
makeshift maintenance chembay, a medbay in disrepair, a
bar/kitchen/botany, a destroyed comms array, a small cargo bay, and a
pretty massive engineering bay. The ship is powered by its own
state-of-the-art "tm" Supermatter.

the ship is as pirate ships are, an admin spawn. Its intent is to serve
as an antagonist in rounds, something that could create some
interesting, and fun roleplay.

the crew:
1x Captain
2x Bridge Officer
1x Head of Security (Master At Arms)
2x Cargo Tech
1x Medical Doctor (Surgeon)
2x Engineer (Carpenter)
1x Atmos Tech
1x Chief Engineer (Head Carpenter)
1x Cook (Steward)
3x Assistant (Deckhand)
3x security officer (Buccaneer)
Ship images:
![2023 02 16-00 52
19](https://user-images.githubusercontent.com/79304582/219223523-45fe98b6-1604-4580-b3a2-0e64280e71a7.png)

![2023 02 16-00 52
22](https://user-images.githubusercontent.com/79304582/219223531-df49a674-465b-4007-8bf1-247134ca4b19.png)


![image](https://user-images.githubusercontent.com/79304582/219224341-cfd592a1-a9a3-4376-9b26-1357af63cd9a.png)


![image](https://user-images.githubusercontent.com/79304582/219224630-6c7e54e7-5c7b-4b76-9cd9-316641bf54d9.png)
(don't mind the broken table)


![image](https://user-images.githubusercontent.com/79304582/219224885-d30a645b-5979-4544-a3d7-7fb50b5f856b.png)


![image](https://user-images.githubusercontent.com/79304582/219224929-e00bd4c2-5a0e-4782-9587-8fb7f0a39d08.png)


![image](https://user-images.githubusercontent.com/79304582/219224963-42f194e8-a95b-4f01-825f-2dc113203597.png)


![image](https://user-images.githubusercontent.com/79304582/219225013-58121f54-78e0-4e14-ba3d-03b3f2f16ffa.png)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
currently, there isn't much Frontiersmen content, nor pirate content in
the codebase despite them being one of the only real antagonists we
have. This is one of the ships I've made for them, and one I think could
do well in creating some fun RP
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: adds the pirate_cutter
add: adds the scar suit
add: misc frontiersmen jobs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: spockye <79304582+spockye@users.noreply.github.com>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: TheLeopold <kostantin.pavlenko@gmail.com>