Shipturfs and various patchie (#1946)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Makes ship specific turfs so we don't have /plating/ planetary turfs
with infinite atmos.
Replaces planetary and plating turfs on the Glaive and Shepherd.
Swaps the captain's spare the Montagnes starts with for a gold ID.
Fixes a runtime caused by attacking water with anything that isn't a
container.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
We didn't have any proper planetary type turfs that could be used on
ships. This caused infinite atmos issues and a bandaid patch that left
"plating" and "wetplating" that don't have smoothing instead. Also minor
patches for the ships this touches, and a fix to a water tile related
runtime are good for the game.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
add: Added shipturfs
del: Old planetary turf subtypes
fix: Fixed water turfs runtiming when hit with a bat
fix: Replaced tiles on the Glaive and Shepherd
tweak: Swapped the montagnes ID
tweak: Nerfed Glaive mining gear
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>